pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.14",
    "discord-py>=2.5.2",
    "flask>=3.1.1",
    "requests>=2.32.4",
    "spotipy>=2.25.1",
]
